requirements.txt
swarms>=5.6.6,<6.0.0
python-dotenv>=1.0.1,<2.0.0
anthropic>=0.34.1,<0.35.0